+++ src/lib/exporter.ts
@@ -1,7 +1,10 @@
 import { Api } from './api';
 
+export type ExportStructure = 'default' | 'legacy';
+
 export interface ExporterSettings {
   readonly template: string;
+  readonly structure: ExportStructure;
 }
 
 /**
@@ -25,7 +28,7 @@ class Exporter {
     const templatePath = '../templates/' + this.settings.template + '/template.js';
 
     // eslint-disable-next-line @typescript-eslint/no-var-requires
-    require(templatePath).template.call(this, types, dirPath);
+    require(templatePath).template.call(this, types, dirPath, this.settings.structure);
   }
 }