+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.3.0 - 2022-08-23
 
 ### Added
-- New `structure` option for specifying `legacy` nested structure for antora docs, or `default` file structure.
+- New `structure` option for specifying `legacy` nested structure for antora docs.
 
 ## 0.2.1 - 2022-04-29