+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.3.0 - 2022-08-23
+
+### Added
+- New `structure` option for specifying `legacy` nested structure for antora docs.
+
 ## 0.2.1 - 2022-04-29
 
 ### Fixed